migration from Integer id to Long id

+++ src/main/java/com/nlu/tai_lieu_xanh/application/comment/dto/request/CommentUpdateRequest.java
@@ -1,6 +1,6 @@
 package com.nlu.tai_lieu_xanh.application.comment.dto.request;
 
 public record CommentUpdateRequest(
-    Integer commentId,
+    Long commentId,
     String content) {
 }
