+++ src/main/java/com/nlu/tai_lieu_xanh/application/user/dto/response/UserRes.java
@@ -1,10 +1,10 @@
 package com.nlu.tai_lieu_xanh.application.user.dto.response;
 
 public record UserRes(
-    Integer id,
+    Long id,
     String email,
     String fullName,
-    Integer comments,
-    Integer posts,
+    Long comments,
+    Long posts,
     String status) {
 }
