+++ src/main/java/com/nlu/tai_lieu_xanh/infrastructure/messaging/event/mdoc/PreviewEvent.java
@@ -1,7 +1,7 @@
 package com.nlu.tai_lieu_xanh.infrastructure.messaging.event.mdoc;
 
 public record PreviewEvent(
-    Integer MDocId,
+    Long MDocId,
     int numPages,
     String filePath) {
 }
