+++ src/main/java/com/nlu/tai_lieu_xanh/infrastructure/persistence/major/JpaMajorRepository.java
@@ -26,7 +26,7 @@ public class JpaMajorRepository implements MajorRepository {
   }
 
   @Override
-  public Optional<Major> findById(Integer id) {
+  public Optional<Major> findById(Long id) {
     return springDataMajorRepository.findById(id);
   }
