+++ src/main/java/com/nlu/tai_lieu_xanh/infrastructure/persistence/major/SpringDataMajorRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.repository.query.Param;
 
 import com.nlu.tai_lieu_xanh.domain.major.Major;
 
-public interface SpringDataMajorRepository extends JpaRepository<Major, Integer> {
+public interface SpringDataMajorRepository extends JpaRepository<Major, Long> {
   @Query("""
       SELECT m from Major m
       WHERE LOWER(m.name) LIKE LOWER(CONCAT('%',:name,'%'))
