+++ src/main/java/com/nlu/tai_lieu_xanh/infrastructure/persistence/user/SpringDataUserRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.nlu.tai_lieu_xanh.domain.user.User;
 
-interface SpringDataUserRepository extends JpaRepository<User, Integer> {
+interface SpringDataUserRepository extends JpaRepository<User, Long> {
   boolean existsUserByEmail(String emai);
 
   Optional<User> findByEmail(String email);
